react: Split FilterDateRange into separate component
